% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{download_wildfinder}
\alias{download_wildfinder}
\title{Download PanTHERIA dataset}
\usage{
download_wildfinder(file = name_file)
}
\value{
This function returns the path (\code{character}) to the downloaded file
(e.g. \verb{data/pantheria/PanTHERIA_1-0_WR05_Aug2008.txt}).
}
\description{
This function downloads the PanTHERIA dataset (text file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.
The file \verb{PanTHERIA_1-0_WR05_Aug2008.txt} will be stored in
\verb{data/pantheria/}. This folder will be created if required.
}
